+++ lib/src/model/run/response/step_detail.dart
@@ -1,0 +1,21 @@
+import 'message_creation.dart';
+
+class StepDetail {
+  StepDetail({
+    required this.messageCreation,
+    required this.type,
+  });
+
+  MessageCreation messageCreation;
+  String type;
+
+  factory StepDetail.fromJson(Map<String, dynamic> json) => StepDetail(
+        messageCreation: MessageCreation.fromJson(json["message_creation"]),
+        type: json["type"] ?? '',
+      );
+
+  Map<String, dynamic> toJson() => {
+        "message_creation": messageCreation.toJson(),
+        "type": type,
+      };
+}